Grouo pyscript related code into module
